Show in-hand / shadow cards based on facedown prop
Previously, the `revealWhenHiddenTo` and `showHand` properties were
being used to control whether to allow card hover for cards in hand.
However, an easier check is whether the card data sent from the server
is facedown or not - the full data for the card is sent non-facedown in
both cases. Thus, these properties don't need to be used in the client.

Additionally, the sorting for partially revealed hands was broken. This
fixes it by sorting all faceup cards after facedown cards.