Send draw deck visibility to the server
Previously, draw deck visibility was toggled on the server by tracking
when the pile was opened and closed via player clicks. However, this
didn't account for the popup opening and closing without user input due
to search prompts. This is now handled through an onPopupChange event
that is fired whenever the popup visibility has changed - either through
users clicking or from prompts activating. The new visibility is then
sent to the server to update viewing state.