Use earliest released card when pack code missing
Previously, if someone pasted in just "3x Arya Stark" into the deck
editor, the card would not be processed because there are multiple
versions of that character. This would then drop the card entirely,
requiring the user to track down the offending line and manually add the
prefix. The main scenario this would happen is when loading in an older
deck, or when the deck was exported from ThronesDB or CardgameDB before
they added the new versions of a card.

Now, when the deck editor comes across a card that has multiple
versions and does not specify the pack code, it automatically chooses
the earliest version of that card released. So for example, "3x Arya
Stark" will use the Core version of Arya, "2x Mace Tyrell" will use the
House of Thorns version of Mace, and so on.