Remove Rookery from UI
Makes rookery unselectable as an option when creating a game and
removes the rookery section from the deck editor.

This explicitly keeps the Rookery deck prompt and associated socket
messages for now in case there are special occasions we need to
bring this back for, or for re-use in similar deck modification
formats.